Reworked to use ducatonator and added a relic list
Modified the loading process to use the warframe.market ducatonator and item listing. Then to pull up individual results if the item is not in that listing.

Also added a relic listing object that used the Warframe PC Drops page to list out every relic, whether it is vaulted or not, and to list out the plat/ducat values based on warframe.market.

In addition, the relic page allows hiding of relics, which helps slim down the page for readability.

Many other edits, but the major ones have been covered.

Further edits are needed to fully integrate the warframe.market ducatonator into this program.