Automation 1.5/3
Created a passive checker that looks for rewards screens every second, and displays the overlays automatically if it is detected.

Optimization is ongoing, and expected to be necessary.

Can be enabled by a setting in the Settings Window.

Optimized button prompts (not tested)
By using windows global hooks.
Probably won't work since I haven't tested them.

Working on Relic Refinement/Selection overlays that will use the same automation.
Should also add a button prompt for lower end PCs

Will be adding in similar automation for mid mission relic selection, but priority is pre-mission selection.
This is due to issues with OCR accuracy.

Relic names are effectively 2 characters, "A1". Thus, near-perfect accuracy is needed.

May need to create a specialized OCR for this... I hope not.

NOTE: this is not complete. This is only a push due to me losing a week worth of code and not wanting to lose it again. Will be releasing 7.1.0 when all 3 are "complete" or major issues arise with relics.