Tesseract Improvements + Version Out of Date Alert
Changed Tesseract default dictionary to a list of all words used in prime parts.

When the app starts, it now will check the most recent release's tag_name. If that tag doesn't match the current version, a warning sign will show next to the version number and the whole top bar will show a tooltip stating that the version is out of date.

Also, changed the warframe not active checks to more streamlined

Minor change to isRelicWindow to work on my setup. Also detects and stores current Window Style.

Fixed "vaulted" text in the reward window to be not bolded

Updated version to 7.2.4