GetPlayers Update
No longer tried to parse the rarity bar.
Instead uses what we've been setup to parse, the part names.

It looks at 3 locations, the middle of the 1st player part with 4, 3, and 2 players.
It scans the whole height and a small width, equal to 1/2 line height, for any text. Stops on the first spot found.