New Automation Mode
Created new automation mode that detects the window through EE.log parsing

Direct parsing runs into issues with file write/flush delays, that have up to 5s of latency.

This overcomes this by using a "Shared Memory Object" to read the log information directly from the source.

However, this is close to the warframe process, and as such will not be enabled by default and will require an agreement before it is enabled.